+++ canvas_api_token/admin.py
@@ -1,0 +1,6 @@
+
+from django.contrib import admin
+from canvas_api_token.models import CanvasApiToken, CanvasDeveloperKey
+
+admin.site.register(CanvasApiToken)
+admin.site.register(CanvasDeveloperKey)