+++ canvas_api_token/migrations/0001_initial.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name='CanvasApiToken',
+            fields=[
+                ('id', models.AutoField(verbose_name='ID', serialize=False, auto_created=True, primary_key=True)),
+                ('user_id', models.CharField(unique=True, max_length=10)),
+                ('token', models.CharField(max_length=1000)),
+                ('created_on', models.DateTimeField(auto_now_add=True)),
+                ('updated_on', models.DateTimeField(auto_now=True)),
+                ('expires_on', models.DateTimeField(null=True)),
+            ],
+            options={
+                'db_table': 'canvas_api_token',
+            },
+            bases=(models.Model,),
+        ),
+    ]